+++ build.gradle
@@ -11,7 +11,6 @@ repositories {
 dependencies {
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.10.0'
     testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.10.0'
-    testImplementation 'org.mockito:mockito-core:5.12.0'
 }
 
 test {
@@ -29,12 +28,11 @@ test {
 }
 
 application {
-    mainClass.set("duke.Chatnius")
-
+    mainClass.set("seedu.duke.Chatnius")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveFileName = "duke.jar"
     archiveClassifier = null
 }
