--- src/main/java/duke/command/Command.java
@@ -1,18 +1,0 @@
-package duke.command;
-
-import duke.task.Task;
-import duke.task.TaskList;
-import duke.ui.Ui;
-import duke.storage.Storage;
-import duke.exception.DukeException;
-
-
-public abstract class Command {
-
-    public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
-
-    public boolean isExit() {
-        return false;
-    }
-}
-
