--- src/main/java/duke/command/ExitCommand.java
@@ -1,20 +1,0 @@
-package duke.command;
-
-import duke.task.Task;
-import duke.task.TaskList;
-import duke.ui.Ui;
-import duke.storage.Storage;
-import duke.exception.DukeException;
-
-
-public class ExitCommand extends Command {
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showGoodbye(); // prints your quit banner
-    }
-
-    @Override
-    public boolean isExit() {
-        return true; // tells the main loop to stop
-    }
-}
