--- src/main/java/duke/command/ListCommand.java
@@ -1,18 +1,0 @@
-package duke.command;
-
-import duke.task.Task;
-import duke.task.TaskList;
-import duke.ui.Ui;
-import duke.storage.Storage;
-import duke.exception.DukeException;
-
-
-public class ListCommand extends Command {
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.show(tasks.renderList());
-    }
-
-
-}
