--- src/main/java/duke/exception/DukeException.java
@@ -1,13 +1,0 @@
-package duke.exception;
-
-public class DukeException extends Exception {
-    public DukeException(String message) {
-        super(message);
-    }
-    public DukeException(String message, Throwable cause) {
-        super(message, cause);
-    }
-    public DukeException(Throwable cause) {
-        super(cause);
-    }
-}
