--- src/main/java/duke/exception/InvalidTaskIndexException.java
@@ -1,9 +1,0 @@
-package duke.exception;
-
-public class InvalidTaskIndexException extends DukeException {
-
-
-    public InvalidTaskIndexException(int oneBasedIndex) {
-        super("Invalid task number!");
-    }
-}
