--- src/main/java/duke/task/ToDo.java
@@ -1,18 +1,0 @@
-package duke.task;
-
-public class ToDo extends Task {
-
-    public ToDo(String name) {
-        super(name);
-    }
-
-    @Override
-    protected String typeCode() {
-        return "T";
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
