+++ src/main/java/seedu/duke/Chatnius.java
@@ -1,0 +1,53 @@
+package seedu.duke;
+
+import seedu.duke.ui.Ui;
+import seedu.duke.storage.Storage;
+import seedu.duke.task.TaskList;
+import seedu.duke.parser.Parser;
+import seedu.duke.command.Command;
+import seedu.duke.exception.DukeException;
+
+
+public class Chatnius {
+    private final Storage storage;
+    private final TaskList tasks;
+    private final Ui ui;
+
+    public Chatnius(String filePath) {
+        this.ui = new Ui();
+        this.storage = new Storage(filePath);
+        TaskList loaded;
+        try {
+            loaded = new TaskList(storage.load());
+        } catch (DukeException e) {
+            ui.showError(e.getMessage());
+            loaded = new TaskList();
+        }
+        this.tasks = loaded;
+    }
+
+    public void run() {
+        ui.showWelcome();
+        boolean isExit = false;
+        while (!isExit) {
+            try {
+                String fullCommand = ui.readCommand();   // prints your "What you need me for?" prompt
+                ui.showLine();
+                Command c = Parser.parse(fullCommand);
+                c.execute(tasks, ui, storage);
+                isExit = c.isExit();
+            } catch (DukeException e) {
+                ui.showError(e.getMessage());
+            } finally {
+                ui.showLine();
+            }
+        }
+
+        ui.close();
+    }
+
+    public static void main(String[] args) {
+        String filePath = (args.length > 0) ? args[0] : "data/duke.txt";
+        new Chatnius(filePath).run();
+    }
+}
