+++ src/main/java/seedu/duke/command/Command.java
@@ -1,0 +1,17 @@
+package seedu.duke.command;
+
+import seedu.duke.task.TaskList;
+import seedu.duke.ui.Ui;
+import seedu.duke.storage.Storage;
+import seedu.duke.exception.DukeException;
+
+
+public abstract class Command {
+
+    public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
+
+    public boolean isExit() {
+        return false;
+    }
+}
+
