+++ src/main/java/seedu/duke/exception/InvalidTaskIndexException.java
@@ -1,0 +1,9 @@
+package seedu.duke.exception;
+
+public class InvalidTaskIndexException extends DukeException {
+
+
+    public InvalidTaskIndexException(int oneBasedIndex) {
+        super("Invalid task number!");
+    }
+}
